Rename bayer.[ch] to stv0680-bayer.[ch]
To make it obvious that this is specific to the stv0680 camlib,
rename its bayer.[ch] source files to stv0680-bayer.[ch].

This is not strictly necessary, but should help when someone
new to it reads the stv0680 source.
