+++ camlibs/stv0680/stv0680-bayer.h
